backend/requirements.txt
Flask==1.1.0
requests==2.24.0
mysql-connector-python==8.0.21